As per Feature request 3001404 (Panel Editor) Multiple items can be selected, by holding down the ctrl key and clicking on them.
Selected items now have a square box drawn around them.

Positioning can now be adjusted by using the arrow keys, the arrow keys on their own will move the items by 5 pixels either way, by holding shift and then the arrow key, it will move a pixel at a time.

Provides a basic copy & paste function within a single panel, allows for single or multiple items to be copied.

A revised pop-up menu is displayed when multiple items are selected, it only displays the options for Edit Alignment, Remove, Copy.

Should also remove the issue in ubuntu when dragging items on the panel editor, that the popupbox is displayed when it shouldn't be. Nb this needs to be check on a MAC that it doesn't cause a problem.