Provides a "Properties" window for text based items, allows the user to set various text based settings in the popup menu from a single window all at once, (eg Font Size, Color, Background, style, border margins etc) and preview them prior to applying the changes.
Also allows for the user to enter in font sizes that are not listed in the popup.

For texted based sensors, it is possible to preview and configure the text for the four different states.

Initial development of the feature, will need some further refinement, only available with the panel editor, but can be used in the other editors once any issues have been ironed out.

Feature request 2958431 relates.