Update so that panel files are loaded in a logical order no matter what order items appear in the panel file.  This means that sensors and turnouts are loaded first and then they follow in order of that specified in the jmri.Manager.
Many Panel Files written prior to 2.13.1 were not saved in a logical, and items such as logixs would be loaded prior to sensors and turnouts, which it is dependant upon.  Normally this would not be a problem if the logix used systemNames as references, however if userNames were stored then this would cause an error on load.