Initial commit of Virtual Sound Decoder main files (supporting stuff has been previously checked in).
this includes all of the main files for the VSD function.

Basic functionality is here, but it is still quite fragile.
Expect null pointer exceptions and so on.

This code should not have any negative effect unless the
feature is called.

To execute, call the LoadVSDecoderCreationAction object's actionPerformed().

Next checkin will be the change to the DebugMenu to
allow execution from the GUI.