Added the ability to select roster groups from the toolbar
Added the ability to use user defined menu and toolbar xml files if provided

If there is an error with the loading of the preference file then the preference window is now displayed.  If no file can be found then the help window is displayed, if there is a connection issue then just the preference window is displayed.