Provide a BasicRosterEntry interface to allow access to a RosterEntry from the jmri.jmrix package.
Add the ability to create a throttle from a roster entry, this then allows the actions performed on a throttle to directly relate to a specific roster entry.

If a throttle has been created from a RosterEntry, then record the duration that the loco was moving for and the last time the loco was used.  these details are currently stored as an atribute and can be access from Media Pane. Part of feature request 3414943 - keep track of activity