Fix a bug that was preventing the start-up sound from playing at engine start.
The bug was, I was trying to build a queue of buffers
to load up on the Source before playing, but my "queue"
method was clearing the queue with each addition.

So I changed the List to a Queue (backed by 
a LinkedList), removed the clearing function,
and cleaned up (a bit) how the queuing was handled
at the low level.

Meanwhile, I added an Element to the VSD Profile
that indicates which Notch is the "idle" notch,
and therefore is the one that should be played
immediately after the start-up sound.