Integrate into the NetBeans life cycle management structure. Remove unneeded System.in/out/err.close() calls.
Notes:
The @OnStop annotation indicates to the NetBeans Platform that it needs to create an instance of the annotated class whose run() method will be called when the application stops (there are @OnStart and @OnShowing annotations that do the same at application start and after the main window is created as well).
The @ServiceProvider annotation registered an instance of the annotated class with the NetBeans Platform default Lookup object (similar to the JMRI InstanceManager). If a class is annotated with both @OnStop and @ServiceProvider only a single instance is created.