Removed code allowing user to silence restart prompt (since it only appears if its really needed).
AppConfigBundle.properties retains messages that are no longer used. Should they be ignored or removed?

NOTE: There may be instances where clicking Save should prompt to restart, but does not. There may also be instances where clicking Save prompts to restart, but should not. Both cases need to be identified and addressed by updating the isDirty() and isRestartRequired() methods in the Classes that manage the specific preference that should or should not trigger the restart.